AMD-694-fix [frontend] fix: fix css hex check
