+++ frontend/__tests__/fixtures.ts
@@ -1,8 +1,6 @@
 import { CSVReportRequestDTO, ReportRequestDTO } from '@src/clients/report/dto/request';
-import { StyledSpan } from '@src/containers/ReportStep/ReportDetail/style';
 import { ReportResponseDTO } from '@src/clients/report/dto/response';
 import { SOURCE_CONTROL_TYPES } from '@src/constants/resources';
-import React from 'react';
 
 export const PROJECT_NAME = 'Heartbeat';
 export const PROJECT_DESCRIPTION =
@@ -72,7 +70,7 @@ export const DEV_CHANGE_FAILURE_RATE = 'Dev change failure rate';
 export const DEV_MEAN_TIME_TO_RECOVERY = 'Dev mean time to recovery';
 export const REQUIRED_DATA = 'Required metrics';
 export const TEST_PROJECT_NAME = 'test project Name';
-export const ERROR_MESSAGE_COLOR = 'color: rgb(211,47,47)';
+export const ERROR_MESSAGE_COLOR = 'color: #d32f2f';
 export const ERROR_DATE = '02/03/';
 export const CREATE_NEW_PROJECT = 'Create a new project';
 export const IMPORT_PROJECT_FROM_FILE = 'Import project from file';
@@ -526,238 +524,6 @@ export const MOCK_REPORT_RESPONSE: ReportResponseDTO = {
 export const MOCK_RETRIEVE_REPORT_RESPONSE = {
   callbackUrl: 'reports/123',
   interval: 10,
-};
-
-export const EXPECTED_REPORT_VALUES = {
-  velocityList: [
-    { id: 0, name: 'Velocity(Story Point)', valueList: [{ value: 20 }] },
-    { id: 1, name: 'Throughput(Cards Count)', valueList: [{ value: 14 }] },
-  ],
-  cycleTimeList: [
-    {
-      id: 0,
-      name: 'Average cycle time',
-      valueList: [
-        { value: 21.18, unit: '(Days/SP)' },
-        { value: '30.26', unit: '(Days/Card)' },
-      ],
-    },
-    {
-      id: 1,
-      name: 'Total development time / Total cycle time',
-      valueList: [{ value: '57.25%' }],
-    },
-    {
-      id: 2,
-      name: 'Average development time',
-      valueList: [
-        { value: '12.13', unit: '(Days/SP)' },
-        { value: '17.32', unit: '(Days/Card)' },
-      ],
-    },
-  ],
-  classification: [
-    {
-      id: 0,
-      name: 'FS Work Type',
-      valuesList: [{ name: 'Feature Work - Planned', value: '57.14%' }],
-    },
-  ],
-  deploymentFrequencyList: [
-    {
-      id: 0,
-      name: 'fs-platform-onboarding/ :shipit: deploy to PROD',
-      valuesList: [
-        {
-          name: 'Deployment frequency',
-          value: '0.30',
-        },
-      ],
-    },
-    {
-      id: 1,
-      name: 'Average',
-      valuesList: [
-        {
-          name: 'Deployment frequency',
-          value: '0.40',
-        },
-      ],
-    },
-  ],
-  devMeanTimeToRecoveryList: [
-    {
-      id: 0,
-      name: 'Heartbeat/:react: Build Frontend',
-      valuesList: [
-        {
-          name: 'Dev mean time to recovery',
-          value: '4.32',
-        },
-      ],
-    },
-    {
-      id: 1,
-      name: 'Heartbeat/:cloudformation: Deploy infra',
-      valuesList: [
-        {
-          name: 'Dev mean time to recovery',
-          value: '0.00',
-        },
-      ],
-    },
-    {
-      id: 2,
-      name: 'Heartbeat/:rocket: Run e2e',
-      valuesList: [
-        {
-          name: 'Dev mean time to recovery',
-          value: '7.67',
-        },
-      ],
-    },
-    {
-      id: 3,
-      name: 'Average',
-      valuesList: [
-        {
-          name: 'Dev mean time to recovery',
-          value: '4.00',
-        },
-      ],
-    },
-  ],
-  leadTimeForChangesList: [
-    {
-      id: 0,
-      name: 'fs-platform-payment-selector/RECORD RELEASE TO PROD',
-      valuesList: [
-        { name: PR_LEAD_TIME, value: '45.04' },
-        { name: PIPELINE_LEAD_TIME, value: '43.12' },
-        { name: TOTAL_DELAY_TIME, value: '88.17' },
-      ],
-    },
-    {
-      id: 1,
-      name: 'Average',
-      valuesList: [
-        { name: PR_LEAD_TIME, value: '60.79' },
-        { name: PIPELINE_LEAD_TIME, value: '39.03' },
-        { name: TOTAL_DELAY_TIME, value: '99.82' },
-      ],
-    },
-  ],
-  devChangeFailureRateList: [
-    {
-      id: 0,
-      name: 'fs-platform-onboarding/ :shipit: deploy to PROD',
-      valuesList: [
-        {
-          name: 'Dev change failure rate',
-          value: '0.00%(0/2)',
-        },
-      ],
-    },
-    {
-      id: 1,
-      name: 'Average',
-      valuesList: [
-        {
-          name: 'Dev change failure rate',
-          value: '0.00%(0/6)',
-        },
-      ],
-    },
-  ],
-  exportValidityTimeMin: 30,
-  reworkList: [
-    {
-      id: 0,
-      name: <React.Fragment>Total rework</React.Fragment>,
-      valueList: [
-        {
-          value: 111,
-          unit: ' (times)',
-        },
-      ],
-    },
-    {
-      id: 3,
-      name: (
-        <React.Fragment>
-          From <StyledSpan>&apos;block&apos;</StyledSpan> to <StyledSpan>&apos;in dev&apos;</StyledSpan>
-        </React.Fragment>
-      ),
-      valueList: [
-        {
-          value: 111,
-          unit: ' (times)',
-        },
-      ],
-    },
-    {
-      id: 5,
-      name: (
-        <React.Fragment>
-          From <StyledSpan>&apos;review&apos;</StyledSpan> to <StyledSpan>&apos;in dev&apos;</StyledSpan>
-        </React.Fragment>
-      ),
-      valueList: [
-        {
-          value: 111,
-          unit: ' (times)',
-        },
-      ],
-    },
-    {
-      id: 6,
-      name: (
-        <React.Fragment>
-          From <StyledSpan>&apos;waiting for testing&apos;</StyledSpan> to <StyledSpan>&apos;in dev&apos;</StyledSpan>
-        </React.Fragment>
-      ),
-      valueList: [
-        {
-          value: 111,
-          unit: ' (times)',
-        },
-      ],
-    },
-    {
-      id: 8,
-      name: (
-        <React.Fragment>
-          From <StyledSpan>&apos;done&apos;</StyledSpan> to <StyledSpan>&apos;in dev&apos;</StyledSpan>
-        </React.Fragment>
-      ),
-      valueList: [
-        {
-          value: 111,
-          unit: ' (times)',
-        },
-      ],
-    },
-    {
-      id: 9,
-      name: <React.Fragment>Total rework cards</React.Fragment>,
-      valueList: [
-        {
-          value: 111,
-          unit: ' (cards)',
-        },
-      ],
-    },
-    {
-      id: 10,
-      name: <React.Fragment>Rework cards ratio</React.Fragment>,
-      valueList: [
-        {
-          value: '88.88',
-          unit: '% (rework cards/throughput)',
-        },
-      ],
-    },
-  ],
 };
 
 export const EMPTY_REPORT_VALUES: ReportResponseDTO = {
